lecture parts 4 & 5 updated

+++ part4/lecture/controllers/notes.js
@@ -1,5 +1,7 @@
 const notesRouter = require('express').Router()
 const Note = require('../models/note')
+const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 notesRouter.get('/', (request, response) => {
   Note.find({}).then(notes => {
@@ -19,28 +21,33 @@ notesRouter.get('/:id', (request, response, next) => {
     .catch(error => next(error))
 })
 
-notesRouter.post('/', (request, response, next) => {
+notesRouter.post('/',  middleware.userExtractor,async (request, response) => {
   const body = request.body
-
+  if (!request.user) response.status(401).json({ error: 'token missing or invalid' }).end()
+  const user = await User.findById(request.user.id)
   const note = new Note({
     content: body.content,
     important: body.important || false,
-    date: new Date()
+    date: new Date(),
+    user: user.id
   })
+  user.notes = user.notes.concat(note._id)
+  const savedNote = await note.save()
+  await user.save()
+  response.json(savedNote)
 
-  note.save()
-    .then(savedNote => {
-      response.json(savedNote)
-    })
-    .catch(error => next(error))
+
 })
 
-notesRouter.delete('/:id', (request, response, next) => {
-  Note.findByIdAndRemove(request.params.id)
-    .then(() => {
-      response.status(204).end()
-    })
-    .catch(error => next(error))
+notesRouter.delete('/:id',middleware.userExtractor, async (request, response) => {
+  if (!request.user) response.status(401).json({ error: 'token missing or invalid' }).end()
+  const user = await User.findById(request.user.id)
+  if(user.notes.includes(request.params.id) ){
+    await Note.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }else{
+    response.status(403).end()
+  }
 })
 
 notesRouter.put('/:id', (request, response, next) => {
